refactor review, comment

+++ ApiServer/DsFinal/src/main/java/com/doubleslash/fifth/dto/DetailReviewWriteDTO.java
@@ -20,7 +20,7 @@ public class DetailReviewWriteDTO {
 	
 	private String place;
 	
-	private double drink;
+	private float drink;
 	
 	private int hangover;
